Add files via upload
so firstly, our database connection works like this:

$conn = @mysqli_connect($host2, $user, $pwd, $sql_db, $port2);
if (!$conn) { 
  $conn = @mysqli_connect($host, $user, $pwd, $sql_db);
  if (!$conn) {
    die('Database connection failed'); 
  }
}

However, we can't do that. It must look like this:
$conn = @mysqli_connect($host, $user, $pwd, $sql_db);

Therefore, I'm currently changing the code so that it only connects to each person's personal database instead of my database, so everyone needs to download the database into their own machine
secondly, for the about.php page, you remember the contribution section? yeah the information about the each person's contribution must be stored in a database and and must be retrieved from it to display on the about page instead of being displayed statically. aka, u have to use php and mysql to render the contribution section instead of plain html. this is wat im also doing 

[text copied and pasted from my discord message]